Extract Section Heading (WIP) (#5117)
* put button on page

* update preview route

* keep print stuff in intro

* Mdct 2102 - PDF: Preview Page setup (#4920)

* heading

* intro text

* Update McparPdfExport.tsx

* field sections

* tests for new components

* table styling

* dynamic fields table layout

* Update parsing.test.tsx

* updating tests

* moved directories

* Update McparReviewSubmitPage.tsx

* clean it up

* pr requests

* more requests

* even more requests

* alt text fix

* testing fixes

* cleanup


* add dynamic fields (#4958)

* add dynamic fields

* add styling

* PR comment update

* parse field data (#4970)

* parse field data

* choice list support

* parsing tests

* Mdct 2103 (#4977)

* set empty fields to no response text

* set empty fields to no response text

* set empty fields to no response text

* Add recursion for children sections for Export

* MDCT-2122: PDF Metadata (#4988)

* Metadata tags for MCPAR report PDF

* fix unit test

* add print button to review page (#5001)

* add print button to review page

* add print button to review page

* add print button to review page

* drawer form data (#4994)

* drawer form data

* Update ExportedReportSubsection.tsx

* Jest testing (#5028)

* drawerform

* dropdown

* Swap to wrapper for each export section

* Prince UI Integration (#5037)

* prince connection

* unit tests

* coverage

* prettier

* take out console log


* Mdct 2125 (#5078)

* add modalDrawers to pdf

* wip

* wip

* wip

* add z-index

* wip

* do the chakra thing

* wip

* wip

* wip

* ?

* wip

* fix indicators and some styling


* Nested children (#5069)

* save

* nested values in plan drawers

* convert to switch statement

* separated functions

* remove unused variables

* remove unused files

* remove old tests

* Export unit tests (#5099)

* entitycard print version teests

* test status indicators

* Exported Drawer Report Section tests

* ExportedModalDrawerReportSection

* Write tests for and update export section typings

* Remove Section heading for review and submit for export

* parsing

* ExportedStandardReportSection test


* padding for status icons (#5104)

* font size fix

* more parsing tests plus icon file change

* renaming

* extract section heading WIP

* nits (#5112)

* nits part i

* more nits and a bug fix or two

* more testing

* PR feedback

* Update ExportedStandardReportSection.test.tsx

* went with no types

* Update ExportedDrawerReportSection.tsx
